requirements.txt
asgiref==3.5.1
coverage==6.3.2
Django==4.0.4
django-debug-toolbar==3.4.0
django-js-asset==2.0.0
django-mptt==0.13.4
Faker==13.11.0
flake8==4.0.1
mccabe==0.6.1
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
Unidecode==1.3.4
